requirements.txt
python-dotenv
deepagents
langchain-openai
tavily
composio
composio-langchain
discord.py
fastapi
uvicorn
supabase

